move inputsValid to util
