+++ web/modules/util/src/html.ts
@@ -137,6 +137,28 @@ export function hover(element: HTMLElement, opts: HoverOpts) {
   if (opts.move) element.addEventListener("pointermove", opts.move)
 }
 
+/**
+ * Checks if all of the given form inputs are valid via ordinary HTML validation.
+ *
+ * @param inputs - The form inputs to check. Can accept `undefined` or
+ *   `null` values, which will count as "invalid" inputs.
+ */
+export function inputsValid(...inputs: (HTMLInputElement | undefined | null)[]) {
+  for (const input of inputs) {
+    if (
+      !input ||
+      !input.validity.valid ||
+      input.value.length === 0 ||
+      input.disabled ||
+      input.readOnly
+    ) {
+      return false
+    }
+  }
+
+  return true
+}
+
 const OBSERVER_CONFIG = {
   childList: true,
   subtree: true,
